_v0/Docs/Analyzer/requirements.txt
openai>=1.0.0
docling>=2.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
numpy<2.0.0
